Implemented the delete account functionality

+++ app/routes/settings_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, render_template, redirect, url_for, flash
+from flask import Blueprint, render_template, redirect, request, url_for, flash
 from app.forms import ChangePassword
 from flask_login import current_user, login_required, logout_user
 from app import db
@@ -29,13 +29,16 @@ def change_password():
     return render_template("account_managment/change_password.html", form=form)
 
 
-@bp.route("/eliminar-cuenta", methods=["GET"])
+@bp.post("/eliminar_cuenta")
 def delete_account():
-    # CHECK: If this is the correct way to delete a user
-    db.session.delete(current_user)
+    if request.form.get("username") == current_user.username:
 
-    db.session.commit()
+        db.session.delete(current_user)
 
-    logout_user()
+        db.session.commit()
 
-    return redirect(url_for("login_routes.login"))
+        logout_user()
+        flash("Cuenta eliminada.", "success")
+        return redirect(url_for("login_routes.login"))
+    flash("No se pudo eliminar la cuenta. No introdujo el usuario correcto.", "danger")
+    return redirect(url_for("main_routes.my_route"))
